Move all video-section code to video-sections.ts
