+++ web/ts/entry/admins.ts
@@ -12,4 +12,4 @@ export * from "../token-management";
 export * from "../worker";
 export * from "../courseAdminManagement";
 export * from "../notification-management";
-export * from "../VideoSection";
+export { VideoSectionClient, VideoSectionsAdmin } from "../video-sections";
