+++ web/ts/entry/video.ts
@@ -3,4 +3,5 @@ export * from "../watch-admin";
 export * from "../TUMLiveVjs";
 export * from "../watch";
 export * from "../chat";
+export * from "../video-sections";
 // Lecture Units are currently not used, so we don't include them in the bundle at the moment
